rename sourceHas to adapterHas
