+++ src/controller/lib/adapter_has.js
